requirements.txt
tabulate==0.8.9
yahoo-fin==0.8.7
-e git://github.com/szirbikpeti/forked-edited-fbchat-package.git#egg=fbchat
Pillow==8.1.2
forex-python==1.5
GitPython==3.1.14